+++ lib/message_driver.rb
@@ -4,7 +4,7 @@ require 'message_driver/broker'
 require 'message_driver/message'
 require 'message_driver/destination'
 require 'message_driver/adapters/base'
-require 'message_driver/message_sender'
+require 'message_driver/message_publisher'
 require 'message_driver/message_receiver'
 
 module MessageDriver