+++ lib/message_driver/message_publisher.rb
@@ -1,0 +1,7 @@
+module MessageDriver
+  module MessagePublisher
+    def publish(destination, body, headers={}, properties={})
+      Broker.publish(destination, body, headers, properties)
+    end
+  end
+end