+++ lib/message_driver/message_sender.rb
@@ -1,7 +1,0 @@
-module MessageDriver
-  module MessageSender
-    def send_message(destination, body, headers={}, properties={})
-      Broker.send_message(destination, body, headers, properties)
-    end
-  end
-end