+++ spec/message_driver/message_publisher_spec.rb
@@ -3,9 +3,9 @@ require 'spec_helper'
 require 'message_driver/adapters/in_memory_adapter'
 
 module MessageDriver
-  describe MessageDriver::MessageSender do
-    class TestSender
-      include MessageDriver::MessageSender
+  describe MessageDriver::MessagePublisher do
+    class TestPublisher
+      include MessageDriver::MessagePublisher
     end
 
     let(:adapter) { Adapters::InMemoryAdapter.new }
@@ -13,25 +13,25 @@ module MessageDriver
       MessageDriver.configure(adapter: adapter)
     end
 
-    subject { TestSender.new }
+    subject { TestPublisher.new }
 
-    describe "#send_message" do
+    describe "#publish" do
       let(:destination) { "my_queue" }
       let(:body) { "my message body" }
 
       it "only requires destination and body" do
-        Broker.instance.should_receive(:send_message).with(destination, body, {}, {})
+        Broker.instance.should_receive(:publish).with(destination, body, {}, {})
 
-        subject.send_message(destination, body)
+        subject.publish(destination, body)
       end
 
       let(:headers) { {foo: :bar} }
       let(:properties) { {bar: :baz} }
 
       it "also passes through the headers and properties" do
-        Broker.instance.should_receive(:send_message).with(destination, body, headers, properties)
+        Broker.instance.should_receive(:publish).with(destination, body, headers, properties)
 
-        subject.send_message(destination, body, headers, properties)
+        subject.publish(destination, body, headers, properties)
       end
     end
   end